0403 移除歷史殘留程式碼與清理 Fixes #65
