+++ shared/types/dashboard.ts
@@ -1,4 +1,20 @@
-import type { ChartType, ChartDisplayOptions } from "./chart";
+export type ChartType =
+  | "table"
+  | "bar"
+  | "line"
+  | "pie"
+  | "scatter"
+  | "histogram"
+  | "area";
+
+export interface ChartDisplayOptions {
+  colors?: string[];
+  showLegend?: boolean;
+  showTooltip?: boolean;
+  showGrid?: boolean;
+  animation?: boolean;
+  stacking?: boolean;
+}
 
 export type SupportedFileType = "json";
 export type FileType = SupportedFileType | "unknown";
